Restructured directories for package installation
