+++ PyUtils/__init__.py
@@ -1,0 +1,4 @@
+import PyUtils.Tab
+import PyUtils.Timer
+import PyUtils.Debug
+import PyUtils.XMLHeader
