--- __init__.py
@@ -1,4 +1,0 @@
-import Tab
-import Timers
-import Debug
-import XMLHeader
